Remove unrecognized argument
The "--all" argument was previously used when polling the status of BGP sessions on a router. This argument has been removed both in the script and in the documentation but not in the contributed systemd service file, making the service never execute successfully.

Removing the argument makes the script poll the BGP sessions without any errors.